+++ src/main/java/com/hyunmin/jwt/domain/account/service/AccountService.java
@@ -4,8 +4,8 @@ import com.hyunmin.jwt.domain.account.dto.LoginRequestDto;
 import com.hyunmin.jwt.domain.account.dto.LoginResponseDto;
 import com.hyunmin.jwt.domain.account.dto.RegisterRequestDto;
 import com.hyunmin.jwt.domain.account.dto.RegisterResponseDto;
-import com.hyunmin.jwt.domain.account.entity.Member;
-import com.hyunmin.jwt.domain.account.repository.MemberRepository;
+import com.hyunmin.jwt.global.common.entity.Member;
+import com.hyunmin.jwt.global.common.repository.MemberRepository;
 import com.hyunmin.jwt.global.exception.RestException;
 import com.hyunmin.jwt.global.exception.code.ErrorCode;
 import com.hyunmin.jwt.global.security.provider.TokenProvider;
